+++ http/src/test/scala/com/twitter/finatra/http/integration/doeverything/main/DoEverythingServer.scala
@@ -1,7 +1,6 @@
 package com.twitter.finatra.http.integration.doeverything.main
 
 import com.twitter.finagle.Filter
-import com.twitter.finagle.http.filter.AddResponseHeadersFilter
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.filters.CommonFilters
 import com.twitter.finatra.http.integration.doeverything.main.controllers.{ReadHeadersController, DoEverythingController, DoNothingController, NonGuiceController}